# [v3] backport arm64 asm changes.