@lizthegrey 2022-05-10
https://flamegraph.com/share/2ee9b0a9-d089-11ec-a9d4-8ee0130fd632 shows improvement in lz4 time, yay.